refactor(window): icon as AssetHandle<Texture>, title-only, uint32 dims
Reshapes Window per the requested data model:
- icon_width/icon_height/icon_pixels (a decoded buffer carried on the window) ->
  AssetHandle<Texture> icon. The window backend loads the texture and reads its
  pixels when the handle changes (applied_icon_id gates re-apply); apply_settings
  just assigns the handle. update_windows() gains AssetsState& to resolve it.
- name + title collapsed to just title. Cameras target a window by title now
  (renderer + sandbox camera match); the redundant name field is gone.
- width/height are uint32 (casts at the int-typed SDL/gpu boundaries; the
  SDL_GetWindowSizeInPixels out-params go through int temporaries).

Window keeps its blessed backend pImpl (+ move-only special members). No uuid id
was added — with the backend on the window and camera-by-title, nothing needs an
external handle yet (the "use a uuid id" path is for if that changes).

89 tests PASS, toom --selftest PASSED, MSVC clean. NOTE: the SDL icon-apply path
runs only in a windowed session (headless tests don't create a window), so it is
compile- and logic-verified but wants a GUI launch to confirm end-to-end.

+++ engine/src/runtime.cpp
@@ -23,8 +23,8 @@ namespace tbx
         {
             auto window = Window();
             window.title = app.config.title;
-            window.width = app.config.width;
-            window.height = app.config.height;
+            window.width = static_cast<uint32>(app.config.width);
+            window.height = static_cast<uint32>(app.config.height);
             state->windows.open_windows.push_back(std::move(window));
         }
         state->app = std::move(app);
